+++ TheMagic/EpisodeMover.cs
@@ -1,5 +1,7 @@
-﻿using System;
+﻿using Microsoft.VisualBasic.FileIO;
+using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -12,13 +14,15 @@ namespace TheMagic
         {
             None = 0,
             FileDoesNotExist,
-            FileAlreadyExists
+            FileAlreadyExists,
+            CouldNotDeleteDirectory
         }
 
         public void MoveEpisodeFiles(List<VideoFile> episodes)
         {
             foreach (VideoFile episode in episodes)
             {
+
                 if (!File.Exists(episode.SourcePath)) episode.MoveError = MoveErrors.FileDoesNotExist;
                 else
                 {
@@ -29,6 +33,31 @@ namespace TheMagic
                     }
                     else
                         episode.MoveError = MoveErrors.FileAlreadyExists;
+                }
+            }
+
+            DeleteParentDirectories(episodes);
+        }
+
+        private void DeleteParentDirectories(List<VideoFile> episodes)
+        {
+            if (episodes != null && SettingsManager.DeleteParentFolder)
+            {
+                foreach (VideoFile episode in episodes.Where(p => p.MoveError == MoveErrors.None))
+                {
+                    if (!String.IsNullOrEmpty(episode.ParentDirectory) && 
+                        !String.IsNullOrEmpty(episode.ParentDirectoryName) && 
+                        episode.ParentDirectoryName == Path.GetFileNameWithoutExtension(episode.SourcePath))
+                    {
+                        try
+                        {
+                            Directory.Delete(episode.ParentDirectory, true);
+                        }
+                        catch
+                        {
+                            episode.MoveError = MoveErrors.CouldNotDeleteDirectory;
+                        }
+                    }
                 }
             }
         }